webapp_ui_bff/pyproject.toml
[project]
name = "webapp-ui-bff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "msal>=1.32.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "starlette-session>=0.4.3",
    "uvicorn[standard]>=0.34.3",
]
